+++ scooter-rental-core/src/main/java/ru/senla/javacourse/enchilik/scooterrental/core/model/Subscription.java
@@ -30,7 +30,7 @@ public class Subscription {
     private LocalDateTime expirationTime;
 
     @Column(name = "rest_units")
-    private Integer restUnits;
+    private Long restUnits;
 
     public Long getId() {
         return id;
@@ -64,11 +64,11 @@ public class Subscription {
         this.expirationTime = expirationTime;
     }
 
-    public Integer getRestUnits() {
+    public Long getRestUnits() {
         return restUnits;
     }
 
-    public void setRestUnits(Integer restUnits) {
+    public void setRestUnits(Long restUnits) {
         this.restUnits = restUnits;
     }
 }
